GW2Client sends JSON directly; no longer a tuple
Previously, the GW2Client's 'sendNext:' sent a tuple with the JSON, the
response, and the request. This is no longer the case. Now the request
and response info is partially put into 'sendError', and the JSON is the
only thing returned by 'sendNext:'.

Most endpoints have been changed to send their results as a sequence.
Before, the entire JSON blog was returned.